requirements.txt
alembic==1.8.0
bcrypt==3.2.2
blinker==1.4
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-msearch==0.2.9.2
Flask-Reuploaded==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.0
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
pdfkit==1.0.0
Pillow==9.1.1
pycparser==2.21
qrcode==7.3.1
requests==2.28.0
SQLAlchemy==1.4.36
stripe==3.3.0
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
